Move configuration related stuff to configuration namespace
