+++ osu.Game.Rulesets.Hishigata/Configuration/ArrowStyle.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel;
+
+namespace osu.Game.Rulesets.Hishigata.Configuration
+{
+    public enum ArrowStyle
+    {
+        [Description("Sharp")]
+        Sharp,
+
+        [Description("Round")]
+        Round,
+    }
+}
