+++ osu.Game.Rulesets.Hishigata/HishigataRuleset.cs
@@ -10,12 +10,12 @@ using osu.Game.Overlays.Settings;
 using osu.Game.Rulesets.Configuration;
 using osu.Game.Rulesets.Difficulty;
 using osu.Game.Rulesets.Hishigata.Beatmaps;
+using osu.Game.Rulesets.Hishigata.Configuration;
 using osu.Game.Rulesets.Hishigata.Difficulty;
 using osu.Game.Rulesets.Hishigata.Mods;
 using osu.Game.Rulesets.Hishigata.Replays;
 using osu.Game.Rulesets.Hishigata.Scoring;
 using osu.Game.Rulesets.Hishigata.UI;
-using osu.Game.Rulesets.Hishigata.UI.Settings;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Rulesets.Replays.Types;
 using osu.Game.Rulesets.Scoring;
@@ -38,9 +38,9 @@ namespace osu.Game.Rulesets.Hishigata
 
         public override IConvertibleReplayFrame CreateConvertibleReplayFrame() => new HishigataReplayFrame();
 
-        public override RulesetSettingsSubsection CreateSettings () => new HishigataSettingsSubsection(this);
+        public override RulesetSettingsSubsection CreateSettings() => new HishigataSettingsSubsection(this);
 
-        public override IRulesetConfigManager CreateConfig (SettingsStore settings) => new HishigataSettingsManager(settings, RulesetInfo);
+        public override IRulesetConfigManager CreateConfig(SettingsStore settings) => new HishigataRulesetConfigManager(settings, RulesetInfo);
 
         public override IEnumerable<Mod> GetModsFor(ModType type)
         {
