+++ osu.Game.Rulesets.Hishigata/UI/Components/Player.cs
@@ -5,14 +5,12 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.Transforms;
 using osu.Framework.Input.Bindings;
 using osu.Framework.Utils;
 using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Graphics.Containers;
 using osu.Game.Rulesets.Hishigata.Objects.Drawables;
-using osu.Game.Rulesets.Hishigata.UI.Settings;
+using osu.Game.Rulesets.Hishigata.Configuration;
 using osuTK;
 using osuTK.Graphics;
 
@@ -142,7 +140,7 @@ namespace osu.Game.Rulesets.Hishigata.UI.Components
             maskedArrow.Alpha = 0;
             pathArrow.Alpha = 0;
 
-            if (style == ArrowStyle.Pointy)
+            if (style == ArrowStyle.Sharp)
                 maskedArrow.Alpha = 1;
             else
                 pathArrow.Alpha = 1;
@@ -151,9 +149,9 @@ namespace osu.Game.Rulesets.Hishigata.UI.Components
         private Bindable<ArrowStyle> arrowStyleBindable = new Bindable<ArrowStyle>();
 
         [BackgroundDependencyLoader]
-        private void load(HishigataSettingsManager config)
+        private void load(HishigataRulesetConfigManager config)
         {
-            config.BindWith(HishigataSetting.ArrowStyle, arrowStyleBindable);
+            config.BindWith(HishigataRulesetSettings.ArrowStyle, arrowStyleBindable);
             arrowStyleBindable.BindValueChanged(x => setArrowSkin(x.NewValue), true);
         }
     }
