--- osu.Game.Rulesets.Hishigata/UI/Settings/ArrowStyle.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Text;
-
-namespace osu.Game.Rulesets.Hishigata.UI.Settings
-{
-    public enum ArrowStyle
-    {
-        [Description("Round")]
-        Round,
-
-        [Description("Pointy")]
-        Pointy
-    }
-}
