--- osu.Game.Rulesets.Hishigata/UI/Settings/HishigataSetting.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace osu.Game.Rulesets.Hishigata.UI.Settings
-{
-    public enum HishigataSetting
-    {
-        ArrowStyle
-    }
-}
